Implement: - User implemented. Only 3 of 63 tests were failed.

+++ src/main/java/ru/practicum/shareit/user/repository/UserRepository.java
@@ -14,5 +14,5 @@ public interface UserRepository {
 
     List<User> getAll();
 
-    boolean remove(Integer id);
+    boolean remove(Long id);
 }
